agentcoach360/agentcoach360_backend/planning_tools.py
from __future__ import annotations

"""
planning_tools.py

Weekly coaching plan agent used in a loop-style pattern.

The idea:
- Manager or agent asks for a coaching plan.
- This agent turns survey insights + persona into a 7-day plan.
- Users can come back later to refine / extend the plan, which forms
  a loop across multiple turns.
"""

from typing import Optional, Dict, Any

from google.adk.agents.llm_agent import LlmAgent
from google.adk.tools.agent_tool import AgentTool


weekly_plan_agent = LlmAgent(
    model="gemini-2.5-flash",
    name="weekly_plan_agent",
    description=(
        "Creates structured 7-day coaching plans for supervisors or frontline agents, "
        "based on survey insights and prior focus areas."
    ),
    instruction=(
        "You are a coaching planner for a contact center.\n\n"
        "INPUT YOU WILL RECEIVE:\n"
        "- persona: either 'manager' or 'agent'.\n"
        "- identifier: team name (for managers) or agent id (for agents), if available.\n"
        "- insights: JSON with metrics, topics, sample comments (same format as other tools).\n"
        "- optional fields:\n"
        "  * last_focus: the last major focus area from previous sessions.\n"
        "  * last_summary: a short text summary of prior coaching.\n"
        "  * previous_plan_text: existing plan text if the user is asking to refine it.\n\n"
        "YOUR JOB:\n"
        "- Produce a concrete 7-day plan aligned to persona.\n"
        "- For MANAGERS: focus on team-wide actions like huddles, calibration sessions,\n"
        "  reviewing call recordings, tweaking workflows, and reinforcing best practices.\n"
        "- For AGENTS: focus on individual behaviors to practice in each day’s calls.\n"
        "- Use last_focus / last_summary when present to show continuity.\n"
        "- If previous_plan_text is provided, treat this request as a REFINEMENT step:\n"
        "  * Keep what works, adjust what doesn’t.\n"
        "  * Clearly mark what changed.\n\n"
        "OUTPUT FORMAT (ALWAYS):\n"
        "- Short intro paragraph referencing persona and timeframe.\n"
        "- 'Week Overview' section.\n"
        "- 'Day-by-day plan' with entries for Day 1–Day 7.\n"
        "- Optional 'How to review progress at the end of the week' section.\n\n"
        "RULES:\n"
        "- Do not invent metrics beyond what insights provide.\n"
        "- Make actions small, realistic, and observable.\n"
        "- Use bullet points and short sentences so it’s easy to read in a dashboard.\n"
    ),
)

weekly_plan_tool = AgentTool(agent=weekly_plan_agent)
